llpc.h separation
Part of formalizing the split between front-end and middle-end.

llpc.h contained some struct and enum declarations also used by the LGC
(middle-end) interface. This commit separates those out into lgc/Defs.h,
and uses that instead of llpc.h from the middle-end.
