+++ llpc/builder/llpcTargetInfo.h
@@ -30,7 +30,7 @@
  */
 #pragma once
 
-#include "llpc.h"
+#include "lgc/Defs.h"
 #include "llvm/ADT/StringRef.h"
 
 namespace Llpc
