+++ llpc/patch/gfx6/chip/llpcGfx6Chip.h
@@ -33,6 +33,7 @@
 #include <cstdint>
 #include <unordered_map>
 #include "llpcAbiMetadata.h"
+#include "llpcTargetInfo.h"
 
 namespace Llpc
 {
