+++ llpc/patch/gfx9/chip/llpcGfx9Chip.h
@@ -33,6 +33,7 @@
 #include <cstdint>
 #include <unordered_map>
 #include "llpcAbiMetadata.h"
+#include "llpcTargetInfo.h"
 
 namespace Llpc
 {
