+++ llpc/patch/llpcCodeGenManager.h
@@ -34,7 +34,6 @@
 #include "llvm/Support/raw_ostream.h"
 
 #include <string>
-#include "llpc.h"
 
 namespace llvm
 {
