+++ llpc/patch/llpcIntrinsDefs.h
@@ -30,6 +30,8 @@
  */
 #pragma once
 
+#include <stdint.h>
+
 namespace Llpc
 {
